% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiptools-package.R
\docType{package}
\name{chiptools-package}
\alias{chiptools}
\alias{chiptools-package}
\title{chiptools: Miscellaneous ChIP-seq and RNA-seq Processing Functions}
\description{
Analysis helpers accumulated across the GNP/MB cerebellar ChIP-seq and RNA-seq work: mixture-model thresholding of binding signal, upper-tail trimming and segmentation of histone-mark intensities, differential-binding wrappers around DESeq2, count normalisation (RPM/RPKM/TPM), time-course interpolation and half-max timing, GO enrichment wrappers around topGO, and the t-SNE, cluster-line and timeline figures built on top of them. Most functions take a features x samples data frame and are meant to be composed at the console rather than run as a pipeline.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jpurzner/chip_tools}
  \item Report bugs at \url{https://github.com/jpurzner/chip_tools/issues}
}

}
\author{
\strong{Maintainer}: James Purzner \email{jpurzner@gmail.com}

}
\keyword{internal}
