% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook-helpers.R
\name{percentile_ranks}
\alias{percentile_ranks}
\title{Percentile rank of each integer value in a vector}
\usage{
percentile_ranks(vector, from = NA, to = NA, round = 1)
}
\arguments{
\item{vector}{Numeric vector.}

\item{from, to}{Integer range to evaluate over; default to the range of
\code{vector}.}

\item{round}{Decimal places for the returned percentage.}
}
\value{
A data frame of \code{value} and \code{perc.rank}.
}
\description{
For every integer from \code{from} to \code{to}, the percentage of \code{vector} at or
below it. Used in \verb{6_Human_MB.Rmd} to place a mouse-derived score on the
human expression distribution.
}
\examples{
percentile_ranks(c(1, 2, 2, 3, 8))
}
